+++ libs/feature/gFeaturePoint.ts
@@ -40,6 +40,11 @@ export default class PointFeature extends Feature {
     refresh() {
         // 执行坐标转换
         const {x, y, r, sr} = this.shape as IPointShape;
+
+        if (!this.layer?.map) {
+            return;
+        }
+
         const {x: screenX, y: screenY} = this.layer.map.transformGlobalToScreen({x, y});
 
         const dpr = CanvasLayer.dpr;