+++ libs/feature/gFeaturePolyline.ts
@@ -30,6 +30,10 @@ export default class PolylineFeature extends Feature {
     // 执行绘制当前
     // @override
     refresh() {
+        if (!this.layer?.map) {
+            return;
+        }
+
         // 执行坐标转换
         const dpr = CanvasLayer.dpr;
         const scale = this.layer.map.getScale();